requirements.txt
fastapi==0.115.9
langchain
langchain-google-genai
langchain-chroma
mlflow
uvicorn
python-dotenv